+++ src/middlewares/authToken.ts
@@ -1,11 +1,7 @@
-import {
-  NextResponse,
-  type NextFetchEvent,
-  type NextRequest,
-} from "next/server";
+import { NextResponse, type NextFetchEvent, type NextRequest } from 'next/server';
 
-import { CustomMiddleware } from "./chain";
-import { updateSession } from "@/utils/supabase/middleware";
+import { CustomMiddleware } from './chain';
+import { updateSession } from '@/utils/supabase/updateSession';
 
 export function authToken(middleware: CustomMiddleware) {
   return async (request: NextRequest, event: NextFetchEvent) => {
